requirements.txt
apiclient==1.0.4
asgiref==3.8.1
asttokens==3.0.0
attrs==25.1.0
backcall==0.2.0
beautifulsoup4==4.13.0
bleach==6.2.0
certifi==2024.2.2
charset-normalizer==3.4.1
colorama==0.4.6
dateparser==1.2.0
decorator==5.1.1
defusedxml==0.7.1
Django==5.1.2
django-debug-toolbar==4.2.0
django-extensions==3.2.3
django-phonenumber-field==8.0.0
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
docopt==0.6.2
drf-spectacular==0.28.0
drf-yasg==1.21.8
executing==2.2.0
fastjsonschema==2.21.1
graphviz==0.20.3
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
ipython==8.12.3
jedi==0.19.2
Jinja2==3.1.5
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
jupyter_client==8.6.3
jupyter_core==5.7.2
jupyterlab_pygments==0.3.0
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
mistune==3.1.1
nbclient==0.10.2
nbconvert==7.16.6
nbformat==5.10.4
packaging==24.2
pandocfilters==1.5.1
parso==0.8.4
phonenumbers==8.13.54
pickleshare==0.7.5
pipreqs==0.5.0
platformdirs==4.3.6
pluggy==1.5.0
prompt_toolkit==3.0.50
pure_eval==0.2.3
pydotplus==2.0.2
Pygments==2.19.1
PyJWT==2.10.1
pyparsing==3.2.0
pytest==8.3.4
pytest-django==4.9.0
python-dateutil==2.9.0.post0
pytz==2024.2
pywin32==308
PyYAML==6.0.2
pyzmq==26.2.1
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rpds-py==0.22.3
six==1.16.0
soupsieve==2.6
sqlparse==0.5.1
stack-data==0.6.3
tinycss2==1.4.0
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.12.2
tzdata==2024.2
tzlocal==5.2
uritemplate==4.1.1
urllib3==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
yarg==0.1.9
